+++ test-composer-integration/portal-package/src-additional/AdditionalPackage.php
@@ -2,36 +2,20 @@
 
 declare(strict_types=1);
 
-namespace HeptacomFixture\Portal\A;
+namespace HeptacomFixture\Portal\AdditionalPackage;
 
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
-use HeptacomFixture\Portal\A\Dto\ShouldNotBeAService;
-use HeptacomFixture\Portal\AdditionalPackage\AdditionalPackage;
+use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 
-class Portal extends PortalContract
+class AdditionalPackage extends PackageContract
 {
-    public function getContainerExcludedClasses(): array
-    {
-        return \array_merge(parent::getContainerExcludedClasses(), [
-            ShouldNotBeAService::class,
-        ]);
-    }
-
     public function buildContainer(ContainerBuilder $containerBuilder): void
     {
-        parent::buildContainer($containerBuilder);
-
         $this->setSyntheticServices($containerBuilder, [
-            'manual-service-by-portal' => new class() {
+            'manual-service-by-additional-package' => new class() {
             },
         ]);
-    }
-
-    public function getAdditionalPackages(): iterable
-    {
-        yield new AdditionalPackage();
     }
 
     /**